refactor(langy/manager): extract manager.js, thin server.js (rule 11)
Code-review rule 11 (SRP) called the 606-line server.js a SRP
violation — one file owning config, helpers, registry, lifecycle,
SIGTERM, AND the HTTP handler whose inner `req.on("end")` callback
was itself ~210 lines.

Split into:

- `manager.js` (new, ~440 lines, dependency-free except node stdlib):
  - low-level helpers (getFreePort, setupWorkerHome, waitForReadiness,
    createOpenCodeSession, postMessage, streamSessionEvents,
    eventBelongsToSession, TERMINAL_EVENT_TYPES)
  - `makeDefaultStartWorker({sessionsRoot, agentsTemplatePath,
    skillsDir, readinessTimeoutMs})` — production spawn path, returns
    a `startWorker(conversationId, credentials) => WorkerInfo` fn
  - `createWorkerRegistry({startWorker, maxWorkers, idleMs})` —
    returns { workers, getOrSpawnWorker, killWorker, reapIdle,
    shutdownAll }. The race + identity check + reaper-skip-pending
    invariants all live here.
  - `createChatHandler({registry, internalSecret})` — orchestration
    only. The inner callback is decomposed into named helpers:
    `readChatBody` (parse + validate), `acquireTurn` (spawn + mutex),
    `runTurn` (post + stream + session-not-found recovery). The /chat
    handler proper is now ~50 lines of glue, well under the 100-line
    bar.

- `server.js` (rewrite, ~90 lines): reads env, calls the three
  factories, wires up http.createServer, sets up the reaper interval
  + signal handlers, listens. The composition root — no logic.

- `Dockerfile`: COPY manager.js alongside server.js.

Also: `LANGY_REAPER_INTERVAL_MS` is now env-configurable (was
hardcoded at 30s), so smoke tests can exercise the reaper without
sleeping for half a minute.

Verified: 23/23 smoke assertions still pass after the extraction
(9 auth/validation + 14 race/mutex/capacity/spawn-failure). Zero
behavior change.
